Add files via upload
Powerpoint presentation our hackathon group presented to the whole gathering at the end of a full day of hacking! We weren't able to quite get the ResNet results by the 3:30pm deadline to be able to show in our powerpoint but we got some decent results an hour later, just too late to present! I think we have a pretty beneficial use case here for image classification that would be useful for underwater camera trap systems.